OnLandEvent() Triggers Prematurely
By leaving the line `m_Grounded = false;` in the statement `if (m_Grounded && jump)` it will cause in some cases for the `OnLandEvent()` to be triggered prematurely. This is because when `m_Grounded = false;` it will cause `wasGrounded` to be equal to false and therefore trigger the `OnLandEvent()` when `m_Grounded = false;` and therefore `wasGrounded` should actually equal true because it **_was_** grounded previously.